renamed_shared_module
