+++ src/app/components/challenge/challenge.module.ts
@@ -24,7 +24,7 @@ import { ChallengeComponent } from './challenge.component';
 
 // import module
 import { ChallengeRoutingModule } from './challenge-routing.module';
-import { CommonSharedModule } from '../../shared/common-shared.module';
+import { CommonSharedModule } from '../../shared/shared.module';
 import { ChallengelistModule } from '../publiclists/challengelist/challengelist.module';
 
 @NgModule({
