+++ src/app/components/get-involved/get-involved.module.ts
@@ -6,7 +6,7 @@ import { GetInvolvedComponent } from './get-involved.component';
 
 // import module
 import { GetInvolvedRoutingModule } from './get-involved-routing.module';
-import { CommonSharedModule } from '../../shared/common-shared.module';
+import { CommonSharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [
