+++ src/app/components/our-team/our-team.module.ts
@@ -6,7 +6,7 @@ import { OurTeamComponent } from './our-team.component';
 
 // import module
 import { OurTeamRoutingModule } from './our-team-routing.module';
-import { CommonSharedModule } from '../../shared/common-shared.module';
+import { CommonSharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [
