+++ src/app/components/profile/profile.module.ts
@@ -6,7 +6,7 @@ import { ProfileComponent } from './profile.component';
 
 // import module
 import { ProfileRoutingModule } from './profile-routing.module';
-import { CommonSharedModule } from '../../shared/common-shared.module';
+import { CommonSharedModule } from '../../shared/shared.module';
 
 @NgModule({
   declarations: [
