+++ src/app/components/publiclists/challengelist/challengelist.module.ts
@@ -10,7 +10,7 @@ import { TeamcardComponent } from '../teamlist/teamcard/teamcard.component';
 import { CardlistComponent } from '../../utility/cardlist/cardlist.component';
 
 // import module
-import { CommonSharedModule } from '../../../shared/common-shared.module';
+import { CommonSharedModule } from '../../../shared/shared.module';
 
 @NgModule({
   declarations: [
